requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<25.0

# Database
psycopg[binary]>=3.1,<4.0

# Environment
python-dotenv>=1.0,<2.0

# AI (Ollama usa API compatível com OpenAI)
openai>=1.0,<2.0
httpx>=0.27,<1.0

# Dev & Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0
pytest-cov>=4.1,<6.0

# Code quality
black>=24.0,<25.0
isort>=5.13,<6.0
ruff>=0.1,<1.0
